// src/modules/profile/talent/screens/TalentNotificationSettings/types.ts
export interface TalentNotificationSettings{
    oneToOneChats: boolean;
    notificationsOfEventBefore24Hours: boolean;
    groupMessages: boolean;
    applicationAcceptance: boolean;
    calendarItems: boolean;
    locationCheckInCheckOut: boolean;
    receivedFeedback: boolean;
    receivedPayment: boolean;
    reminderMessages24hrs6hrs1hrs: boolean;
    simpleMessages: boolean;
    locationChange: boolean;
  }